- merge() will not issue a needless merge of attributes if the given instance is the same instance which is already present.
- merge() now also merges the "options" associated with a given
state, i.e. those passed through query.options() which follow
along with an instance, such as options to eagerly- or
lazyily- load various attributes.   This is essential for
the construction of highly integrated caching schemes.  This
is a subtle behavioral change vs. 0.5.

- A bug was fixed regarding the serialization of the "loader
path" present on an instance's state, which is also necessary
when combining the usage of merge() with serialized state
and associated options that should be preserved.  

- The "query_cache" examples have been removed, and are replaced
with a fully comprehensive approach that combines the usage of
Beaker with SQLAlchemy.  New query options are used to indicate
the caching characteristics of a particular Query, which 
can also be invoked deep within an object graph when lazily
loading related objects.  See /examples/beaker_caching/README.



git-svn-id: http://svn.sqlalchemy.org/sqlalchemy/trunk@6625 8cd8332f-0806-0410-a4b6-96f4b9520244